fix(web-app): sonarcloud issue, offload logic to github file
